% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateProteinInfoFromIndra.R
\name{.populateHgncIdsInDataFrame}
\alias{.populateHgncIdsInDataFrame}
\title{Populate HGNC IDs in Data Frame}
\usage{
.populateHgncIdsInDataFrame(df)
}
\arguments{
\item{df}{A data frame containing protein information.}
}
\value{
A data frame with populated HGNC IDs.
}
\description{
This function populates the HGNC IDs in the data frame based on the Uniprot IDs.
}
