% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateProteinInfoFromIndra.R
\name{.populateHgncNamesInDataFrame}
\alias{.populateHgncNamesInDataFrame}
\title{Populate HGNC Names in Data Frame}
\usage{
.populateHgncNamesInDataFrame(df)
}
\arguments{
\item{df}{A data frame containing protein information.}
}
\value{
A data frame with populated HGNC names.
}
\description{
This function populates the HGNC names in the data frame based on the HGNC IDs.
}
