% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateProteinInfoFromIndra.R
\name{.populateKinaseInfoInDataFrame}
\alias{.populateKinaseInfoInDataFrame}
\title{Populate Kinase Info in Data Frame}
\usage{
.populateKinaseInfoInDataFrame(df)
}
\arguments{
\item{df}{A data frame containing protein information.}
}
\value{
A data frame with populated kinase information.
}
\description{
This function populates the kinase information in the data frame based on the HGNC names.
}
