% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateProteinInfoFromIndra.R
\name{.populatePhophataseInfoInDataFrame}
\alias{.populatePhophataseInfoInDataFrame}
\title{Populate Phosphatase Info in Data Frame}
\usage{
.populatePhophataseInfoInDataFrame(df)
}
\arguments{
\item{df}{A data frame containing protein information.}
}
\value{
A data frame with populated phosphatase information.
}
\description{
This function populates the phosphatase information in the data frame based on the HGNC names.
}
