% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateProteinInfoFromIndra.R
\name{.populateTranscriptionFactorInfoInDataFrame}
\alias{.populateTranscriptionFactorInfoInDataFrame}
\title{Populate Transcription Factor Info in Data Frame}
\usage{
.populateTranscriptionFactorInfoInDataFrame(df)
}
\arguments{
\item{df}{A data frame containing protein information.}
}
\value{
A data frame with populated transcription factor information.
}
\description{
This function populates the transcription factor information in the data frame based on the HGNC names.
}
