% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateProteinInfoFromIndra.R
\name{.populateUniprotIdsInDataFrame}
\alias{.populateUniprotIdsInDataFrame}
\title{Populate Uniprot IDs in Data Frame}
\usage{
.populateUniprotIdsInDataFrame(df, proteinIdType)
}
\arguments{
\item{df}{A data frame containing protein information.}

\item{proteinIdType}{A character string specifying the type of protein ID. 
It can be either "Uniprot" or "Uniprot_Mnemonic".}
}
\value{
A data frame with populated Uniprot IDs.
}
\description{
This function populates the Uniprot IDs in the data frame based on the protein ID type.
}
