% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateProteinInfoFromIndra.R
\name{.validateAnnotateProteinInfoFromIndraInput}
\alias{.validateAnnotateProteinInfoFromIndraInput}
\title{Validate Annotate Protein Info Input}
\usage{
.validateAnnotateProteinInfoFromIndraInput(df)
}
\arguments{
\item{df}{A data frame containing protein information.}
}
\value{
None. Throws an error if validation fails.
}
\description{
This function validates the input data frame for the annotateProteinInfoFromIndra function.
}
