% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizeSubnetwork.R
\name{visualizeSubnetwork}
\alias{visualizeSubnetwork}
\title{Create visualization of subnetwork in cytoscape}
\usage{
visualizeSubnetwork(nodes, edges, pvalue_cutoff = 0.05, logfc_cutoff = 0.5)
}
\arguments{
\item{nodes}{dataframe of nodes}

\item{edges}{dataframe of edges}

\item{pvalue_cutoff}{p-value cutoff for coloring significant proteins. Default is 0.05}

\item{logfc_cutoff}{log fold change cutoff for coloring significant proteins. Default is 0.5}
}
\description{
Create visualization of subnetwork in cytoscape
}
\examples{
input = data.table::fread(system.file("processed_data/groupComparisonModel.csv", 
                             package = "MSstatsPPI"))
# subnetwork = getSubnetworkFromIndra(input)
# visualizeSubnetwork(subnetwork$nodes, subnetwork$edges)

              
}
