% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSubnetworkFromIndra.R
\name{getSubnetworkFromIndra}
\alias{getSubnetworkFromIndra}
\title{Get subnetwork from INDRA database}
\usage{
getSubnetworkFromIndra(input, pvalueCutoff = NULL)
}
\arguments{
\item{input}{output of \code{\link[MSstats]{groupComparison}} function's 
comparisionResult table, which contains a list of proteins and their 
corresponding p-values, logFCs, along with additional HGNC ID and HGNC 
name columns}

\item{pvalueCutoff}{p-value cutoff for filtering. Default is NULL, i.e. no
filtering}
}
\value{
list of 2 data.frames, nodes and edges
}
\description{
Using differential abundance results from MSstats, this function retrieves
a subnetwork of protein interactions from INDRA database.
}
\examples{
input <- data.table::fread(system.file(
    "extdata/groupComparisonModel.csv",
    package = "MSstatsBioNet"
))
subnetwork <- getSubnetworkFromIndra(input, pvalueCutoff = 0.05)
head(subnetwork$nodes)
head(subnetwork$edges)

}
